+++ R/globals.R
@@ -1,3 +1,3 @@
 utils::globalVariables(
-  c("%>%", ".", "assessment_description", "curr_grade_lvl", "discard", "discard1", "discard2", "report_school_year", "standard_achieved", "start_year", "test_subject", "test_year")
+  c("%>%", ".", "assessment_description", "curr_grade_lvl", "discard", "discard1", "discard2", "end_performance_level_numeric", "end_scale_score", "end_test_year", "ethnic_desc", "midpoint", "most_recent", "n", "n_level", "n_total", "numeric_score", "pct", "performance_level", "performance_level_numeric", "proficient_dummy", "report_school_year", "scale_score", "standard_achieved", "start_performance_level_numeric", "start_test_year", "start_scale_score", "start_year", "student_gender", "student_id", "test_grade",  "test_subject", "test_year")
 )